+++ packages/language-server-ruby/src/formatters/RubyFMT.ts
@@ -1,15 +1,14 @@
-import path from 'path';
+import { URI } from 'vscode-uri'
 import BaseFormatter from './BaseFormatter';
-
-const RUBYFMT_PATH = path.resolve(__dirname, 'rubyfmt.rb');
 
 export default class RubyFMT extends BaseFormatter {
 	get cmd(): string {
-		return 'ruby';
+		return 'rubyfmt';
 	}
 
 	get args(): string[] {
-		return ['--disable=gems', RUBYFMT_PATH];
+		const documentPath = URI.parse(this.document.uri);
+		return [documentPath.fsPath];
 	}
 
 	get useBundler(): boolean {